requirements.txt
astroid==2.2.5
attrs==19.3.0
Automat==0.8.0
autopep8==1.4.4
beautifulsoup4==4.8.2
boto3==1.11.16
botocore==1.14.16
bs4==0.0.1
certifi==2019.6.16
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.4
django-allauth==0.39.1
django-cors-headers==3.0.2
django-countries==5.3.3
django-rest-auth==0.9.5
django-storages==1.9.1
djangorestframework==3.9.4
docutils==0.15.2
gunicorn==19.9.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.21
jmespath==0.9.4
lazy-object-proxy==1.4.1
lxml==4.5.0
mccabe==0.6.1
oauthlib==3.0.1
parsel==1.5.2
pep8==1.7.1
Pillow==6.1.0
# pkg-resources==0.0.0
Protego==0.1.16
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==2.0.0
pylint==2.3.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-decouple==3.1
python-scrapyd-api==2.1.2
python3-openid==3.1.0
pytz==2019.1
queuelib==1.5.0
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.3.3
Scrapy==1.8.0
scrapyd==1.2.1
service-identity==18.1.0
six==1.12.0
soupsieve==1.9.5
sqlparse==0.3.0
stripe==2.33.0
Twisted==19.10.0
typed-ast==1.4.0
urllib3==1.25.3
w3lib==1.21.0
whitenoise==5.0.1
wrapt==1.11.2
zope.interface==4.7.1
